backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Longwei Su",email = "longwei.su@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "flask-httpauth (>=4.8.0,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)"
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
